bugfix: do not show token supply
