+++ app/(tabs)/Price.tsx
