+++ app/(tabs)/SupplyDynamics.tsx
