+++ app/(tabs)/TokenDetails.tsx
