+++ hardhat/ignition/modules/MyToken.js
@@ -1,10 +1,14 @@
 const { buildModule } = require("@nomicfoundation/hardhat-ignition/modules");
+const { default: BigNumber } = require("bignumber.js");
 
 module.exports = buildModule("MyTokenModule", (m) => {
   const name = m.getParameter("name", process.env.NAME);
   const symbol = m.getParameter("symbol", process.env.SYMBOL);
-  const decimals = m.getParameter("decimals", process.env.DECIMALS);
-  const supply = m.getParameter("supply", process.env.SUPPLY);
-  const myToken = m.contract("MyToken", [name, symbol, decimals, supply]);
+  const supply = m.getParameter(
+    "supply",
+    BigNumber(process.env.SUPPLY).times(BigNumber(10).pow(18)).toString(10)
+  );
+
+  const myToken = m.contract("MyToken", [name, symbol, supply]);
   return { myToken };
 });
